Convert project metadata to the declarative format
The effects of this on the metadata in distribution packages are as follows:

* The deprecated `Home-page` metadata field is replaced with
  `Project-URL: Homepage, ...`.

* The deprecated `License` metadata field and
  `License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)` classifier
  are replaced with the new `License-Expression` field.

* The `Author` field is merged into `Author-email`.

* The `Author-email`, `Classifier`, `Description`, `Project-URL`,
  `Requires-Python`, and `Summary` fields are no longer marked dynamic.

With newer versions of setuptools, all the files listed in `MANIFEST.in` are
already included by default, so remove it.

Note that setuptools no longer depends on certifi, so the comment in `setup.py`
is no longer relevant.